cmake: Remove AI library
