+++ libs/CMakeLists.txt
@@ -2,7 +2,6 @@
 # Libaries cmake file.
 # --------------------------------------------------------------------------------------------------
 
-add_subdirectory(ai)
 add_subdirectory(app)
 add_subdirectory(asset)
 add_subdirectory(check)
