--- libs/ai/include/ai_result.h
@@ -1,10 +1,0 @@
-#pragma once
-#include "core_string.h"
-
-typedef enum {
-  AiResult_Running,
-  AiResult_Success,
-  AiResult_Failure,
-} AiResult;
-
-String ai_result_str(AiResult);
