--- libs/ai/src/node/failure.c
@@ -1,13 +1,0 @@
-#include "ai_eval.h"
-#include "asset_behavior.h"
-#include "core_diag.h"
-
-AiResult ai_node_failure_eval(const AiEvalContext* ctx, const AssetAiNodeId nodeId) {
-  const AssetAiNode* def = &ctx->nodeDefs[nodeId];
-  diag_assert(def->type == AssetAiNode_Failure);
-
-  (void)ctx;
-  (void)def;
-
-  return AiResult_Failure;
-}
