--- libs/ai/src/node/success.c
@@ -1,13 +1,0 @@
-#include "ai_eval.h"
-#include "asset_behavior.h"
-#include "core_diag.h"
-
-AiResult ai_node_success_eval(const AiEvalContext* ctx, const AssetAiNodeId nodeId) {
-  const AssetAiNode* def = &ctx->nodeDefs[nodeId];
-  diag_assert(def->type == AssetAiNode_Success);
-
-  (void)ctx;
-  (void)def;
-
-  return AiResult_Success;
-}
