--- libs/ai/test/config.c
@@ -1,16 +1,0 @@
-#include "app_check.h"
-
-void app_check_configure(CheckDef* check) {
-  register_spec(check, node_condition);
-  register_spec(check, node_execute);
-  register_spec(check, node_failure);
-  register_spec(check, node_invert);
-  register_spec(check, node_parallel);
-  register_spec(check, node_repeat);
-  register_spec(check, node_running);
-  register_spec(check, node_selector);
-  register_spec(check, node_sequence);
-  register_spec(check, node_success);
-  register_spec(check, node_try);
-  register_spec(check, tracer_record);
-}
