--- libs/debug/include/debug_brain.h
@@ -1,5 +1,0 @@
-#pragma once
-#include "ecs_entity.h"
-#include "ecs_module.h"
-
-EcsEntityId debug_brain_panel_open(EcsWorld*, EcsEntityId window);
