--- libs/scene/src/sensor.c
@@ -1,137 +1,0 @@
-#include "core_math.h"
-#include "core_stringtable.h"
-#include "ecs_world.h"
-#include "scene_attack.h"
-#include "scene_brain.h"
-#include "scene_faction.h"
-#include "scene_health.h"
-#include "scene_knowledge.h"
-#include "scene_nav.h"
-#include "scene_register.h"
-#include "scene_target.h"
-#include "scene_time.h"
-#include "scene_transform.h"
-
-// clang-format off
-
-static StringHash g_brainKeyTime,
-                  g_brainKeyEntity,
-                  g_brainKeyPosition,
-                  g_brainKeyHealth,
-                  g_brainKeyFaction,
-                  g_brainKeyNavArrived,
-                  g_brainKeyAttacking,
-                  g_brainKeyTargetRangeMax,
-                  g_brainKeyTargetEntity,
-                  g_brainKeyTargetPos,
-                  g_brainKeyTargetDist,
-                  g_brainKeyTargetLos;
-
-// clang-format on
-
-ecs_view_define(SensorGlobalView) { ecs_access_read(SceneTimeComp); }
-
-ecs_view_define(BrainView) {
-  ecs_access_read(SceneBrainComp);
-  ecs_access_write(SceneKnowledgeComp);
-  ecs_access_maybe_read(SceneAttackComp);
-  ecs_access_maybe_read(SceneFactionComp);
-  ecs_access_maybe_read(SceneHealthComp);
-  ecs_access_maybe_read(SceneNavAgentComp);
-  ecs_access_maybe_read(SceneTargetFinderComp);
-  ecs_access_maybe_read(SceneTransformComp);
-}
-
-ecs_system_define(SceneSensorUpdateSys) {
-  EcsView*     globalView = ecs_world_view_t(world, SensorGlobalView);
-  EcsIterator* globalItr  = ecs_view_maybe_at(globalView, ecs_world_global(world));
-  if (!globalItr) {
-    return;
-  }
-  const SceneTimeComp* timeComp = ecs_view_read_t(globalItr, SceneTimeComp);
-
-  EcsView* view = ecs_world_view_t(world, BrainView);
-  for (EcsIterator* itr = ecs_view_itr_step(view, parCount, parIndex); ecs_view_walk(itr);) {
-    const EcsEntityId     entity = ecs_view_entity(itr);
-    const SceneBrainComp* brain  = ecs_view_read_t(itr, SceneBrainComp);
-    if (scene_brain_flags(brain) & SceneBrainFlags_PauseSensors) {
-      continue;
-    }
-    SceneKnowledgeComp* knowledge = ecs_view_write_t(itr, SceneKnowledgeComp);
-
-    scene_knowledge_set(knowledge, g_brainKeyTime, script_time(timeComp->time));
-    scene_knowledge_set(knowledge, g_brainKeyEntity, script_entity(entity));
-
-    const SceneTransformComp* transform = ecs_view_read_t(itr, SceneTransformComp);
-    if (transform) {
-      scene_knowledge_set(knowledge, g_brainKeyPosition, script_vec3(transform->position));
-    }
-
-    const SceneHealthComp* health = ecs_view_read_t(itr, SceneHealthComp);
-    if (health) {
-      scene_knowledge_set(knowledge, g_brainKeyHealth, script_num(health->norm));
-    }
-
-    const SceneFactionComp* faction = ecs_view_read_t(itr, SceneFactionComp);
-    if (faction) {
-      scene_knowledge_set(knowledge, g_brainKeyFaction, script_num(faction->id));
-    }
-
-    const SceneNavAgentComp* navAgent = ecs_view_read_t(itr, SceneNavAgentComp);
-    if (navAgent) {
-      if (navAgent->flags & SceneNavAgent_Traveling) {
-        scene_knowledge_set_null(knowledge, g_brainKeyNavArrived);
-      } else {
-        scene_knowledge_set(knowledge, g_brainKeyNavArrived, script_vec3(navAgent->targetPos));
-      }
-    }
-
-    const SceneAttackComp* attack = ecs_view_read_t(itr, SceneAttackComp);
-    if (attack) {
-      const bool isAttacking = (attack->flags & SceneAttackFlags_Firing) != 0;
-      scene_knowledge_set(knowledge, g_brainKeyAttacking, script_bool(isAttacking));
-    }
-
-    const SceneTargetFinderComp* targetFinder = ecs_view_read_t(itr, SceneTargetFinderComp);
-    const EcsEntityId            target = targetFinder ? scene_target_primary(targetFinder) : 0;
-    if (target) {
-      const bool los = (targetFinder->flags & SceneTarget_LineOfSight) != 0;
-
-      scene_knowledge_set(knowledge, g_brainKeyTargetRangeMax, script_num(targetFinder->rangeMax));
-      scene_knowledge_set(knowledge, g_brainKeyTargetEntity, script_entity(target));
-      scene_knowledge_set(
-          knowledge, g_brainKeyTargetPos, script_vec3(targetFinder->targetPosition));
-      scene_knowledge_set(
-          knowledge, g_brainKeyTargetDist, script_num(targetFinder->targetDistance));
-      scene_knowledge_set(knowledge, g_brainKeyTargetLos, script_bool(los));
-    } else {
-      scene_knowledge_set_null(knowledge, g_brainKeyTargetEntity);
-      scene_knowledge_set_null(knowledge, g_brainKeyTargetPos);
-      scene_knowledge_set_null(knowledge, g_brainKeyTargetDist);
-      scene_knowledge_set_null(knowledge, g_brainKeyTargetLos);
-    }
-  }
-}
-
-ecs_module_init(scene_sensor_module) {
-  g_brainKeyTime           = stringtable_add(g_stringtable, string_lit("global_time"));
-  g_brainKeyEntity         = stringtable_add(g_stringtable, string_lit("self_entity"));
-  g_brainKeyPosition       = stringtable_add(g_stringtable, string_lit("self_position"));
-  g_brainKeyHealth         = stringtable_add(g_stringtable, string_lit("self_health"));
-  g_brainKeyFaction        = stringtable_add(g_stringtable, string_lit("self_faction"));
-  g_brainKeyNavArrived     = stringtable_add(g_stringtable, string_lit("self_nav_arrived"));
-  g_brainKeyAttacking      = stringtable_add(g_stringtable, string_lit("self_attacking"));
-  g_brainKeyTargetRangeMax = stringtable_add(g_stringtable, string_lit("target_range_max"));
-  g_brainKeyTargetEntity   = stringtable_add(g_stringtable, string_lit("target_entity"));
-  g_brainKeyTargetPos      = stringtable_add(g_stringtable, string_lit("target_position"));
-  g_brainKeyTargetDist     = stringtable_add(g_stringtable, string_lit("target_distance"));
-  g_brainKeyTargetLos      = stringtable_add(g_stringtable, string_lit("target_los"));
-
-  ecs_register_view(SensorGlobalView);
-  ecs_register_view(BrainView);
-
-  ecs_register_system(SceneSensorUpdateSys, ecs_view_id(SensorGlobalView), ecs_view_id(BrainView));
-
-  ecs_order(SceneSensorUpdateSys, SceneOrder_SensorUpdate);
-  ecs_parallel(SceneSensorUpdateSys, 2);
-}
