+++ libs/scene/test/config.c
@@ -1,6 +1,6 @@
 #include "app_check.h"
 
 void app_check_configure(CheckDef* check) {
-  register_spec(check, brain);
   register_spec(check, nav);
+  register_spec(check, script);
 }
